+++ nosku.R
@@ -2,10 +2,6 @@
 if(!is.null(reporte)){
   if(unique(reporte$Estado) == "El SKU no se encuentra en cubo"){
     
-    setwd("S:/OMNI/ReporteGeneral")
-    
-    csv<-read.csv("ProblemasSKU.csv")
-    
     Producto<-reporte$Archivo
     
     Responsable <- as.character(Sys.info()["user"])
@@ -14,9 +10,42 @@ if(!is.null(reporte)){
     
     nosku<-data.frame("NoSKUCubo", Producto, Responsable, Fecha)
     
-    nosku<-rbind(csv, nosku)
+    if(dir.exists("S:/OMNI/ReporteGeneral")){
+      setwd("S:/OMNI/ReporteGeneral")
+      
+      csv<-read.csv("ProblemasSKU.csv")
+      
+      nosku<-rbind(csv, nosku)
+      
+      write.csv(nosku, "ProblemasSKU.csv", row.names = FALSE)
+      
+    }else{
+      
+      setwd(guardar)
+      
+      if (file.exists("ENVIAR-a-RESPONSABLE-ProblemasSKU.csv")){
+        
+        
+        csv<-read.csv("ENVIAR-a-RESPONSABLE-ProblemasSKU.csv")
+        csv<-data.frame(csv)
+        
+        csv<-rbind(csv, nosku)
+        
+        write.csv(csv, "ENVIAR-a-RESPONSABLE-ProblemasSKU.csv", row.names = FALSE)
+        
+        
+      }else{
+        write.csv(nosku, "ENVIAR-a-RESPONSABLE-ProblemasSKU.csv", row.names = FALSE)
+        
+        
+      }
+      
+      
+      
+      
+    }
     
-    write.csv(nosku, "ProblemasSKU.csv", row.names = FALSE)
+    
     
   }
